refactor: extract, restructure and reuse types, utils, components...
