+++ packages/plugins/text/src/types/config.ts
@@ -1,0 +1,131 @@
+import { EditorPluginProps, SerializedScalarStateType } from '@edtr-io/plugin'
+import { DeepPartial } from '@edtr-io/ui'
+import { Descendant, Range } from 'slate'
+
+/** @public */
+export type TextPluginState = SerializedScalarStateType<
+  Descendant[],
+  { value: Descendant[]; selection: Range | null }
+>
+
+/** @public */
+export type TextProps = EditorPluginProps<TextPluginState, TextConfig>
+
+/** @public */
+export interface TextConfig {
+  placeholder?: TextPluginConfig['placeholder']
+  plugins?: TextConfigPlugins
+  /** @deprecated */
+  registry: TextPluginConfig['registry']
+  i18n?: DeepPartial<TextPluginConfig['i18n']>
+  theme?: DeepPartial<TextPluginConfig['theme']>
+  blockquote?: string
+  noLinebreaks?: boolean
+}
+
+/** @public */
+export interface TextConfigPlugins {
+  code?: boolean
+  colors?: boolean
+  headings?: boolean
+  katex?: boolean
+  links?: boolean
+  lists?: boolean
+  math?: boolean
+  paragraphs?: boolean
+  richText?: boolean
+  suggestions?: boolean
+}
+
+/** @public */
+export interface TextPluginConfig {
+  placeholder: string
+  enabledPlugins: TextConfigPlugins
+  // TODO
+  //plugins: ((pluginClosure: SlatePluginClosure) => TextPlugin)[]
+  registry: {
+    name: string
+    title?: string
+    description?: string
+  }[]
+  i18n: {
+    blockquote: {
+      toggleTitle: string
+    }
+    code: {
+      toggleTitle: string
+    }
+    colors: {
+      setColorTitle: string
+      resetColorTitle: string
+      openMenuTitle: string
+      closeMenuTitle: string
+    }
+    headings: {
+      setHeadingTitle(level: number): string
+      openMenuTitle: string
+      closeMenuTitle: string
+    }
+    link: {
+      toggleTitle: string
+      placeholder: string
+      openInNewTabTitle: string
+    }
+    list: {
+      toggleOrderedList: string
+      toggleUnorderedList: string
+      openMenuTitle: string
+      closeMenuTitle: string
+    }
+    math: {
+      toggleTitle: string
+      displayBlockLabel: string
+      placeholder: string
+      editors: {
+        visual: string
+        latex: string
+        noVisualEditorAvailableMessage: string
+      }
+      helpText(
+        KeySpan: React.ComponentType<{ children: React.ReactNode }>
+      ): React.ReactNode
+    }
+    richText: {
+      toggleStrongTitle: string
+      toggleEmphasizeTitle: string
+    }
+    suggestions: {
+      noResultsMessage: string
+    }
+  }
+  theme: {
+    backgroundColor: string
+    color: string
+    hoverColor: string
+    active: {
+      backgroundColor: string
+      color: string
+    }
+    dropDown: {
+      backgroundColor: string
+    }
+    suggestions: {
+      background: {
+        default: string
+        highlight: string
+      }
+      text: {
+        default: string
+        highlight: string
+      }
+    }
+    plugins: {
+      colors: {
+        colors: string[]
+        defaultColor: string
+      }
+    }
+  }
+  blockquote?: string
+  noLinebreaks?: boolean
+}
