+++ packages/plugins/text/src/types/index.ts
@@ -1,0 +1,4 @@
+export * from './config'
+export * from './hovering-toolbar'
+export * from './math-element'
+export * from './text-editor'
