+++ packages/plugins/text/src/types/math-element.ts
@@ -1,0 +1,12 @@
+// TODO: Good structure
+export interface MathElement {
+  type: 'math'
+  src: string
+  inline: boolean
+  children: []
+}
+
+export interface TransformOutOfElementProps {
+  reverse?: boolean
+  shouldDelete?: boolean
+}
