+++ packages/plugins/text/src/utils/document.ts
@@ -1,0 +1,9 @@
+import { StateTypeValueType } from '@edtr-io/plugin'
+
+import { TextPluginState } from '../types'
+
+export const emptyDocumentFactory =
+  (): StateTypeValueType<TextPluginState> => ({
+    value: [{ type: 'p', children: [{ text: '' }] }],
+    selection: null,
+  })
