+++ packages/plugins/text/src/utils/is-touch-device.ts
@@ -1,0 +1,9 @@
+export function isTouchDevice() {
+  return (
+    typeof window !== 'undefined' &&
+    ('ontouchstart' in window ||
+      navigator.maxTouchPoints > 0 ||
+      // @ts-expect-error TODO:
+      navigator.msMaxTouchPoints > 0)
+  )
+}
