+++ packages/plugins/text/src/utils/with-lists-plugin.ts
@@ -1,0 +1,56 @@
+import { ListType, withLists } from '@prezly/slate-lists'
+import { Element, Node } from 'slate'
+
+import {
+  ListItem,
+  ListItemText,
+  OrderedList,
+  Paragraph,
+  UnorderedList,
+} from '../types'
+
+// TODO: Use enum for types here as in https://www.npmjs.com/package/@prezly/slate-lists
+// TODO: Fix "as ..." parts in the functions
+export const withListsPlugin = withLists({
+  isConvertibleToListTextNode(node: Node) {
+    return Element.isElementType(node, 'p')
+  },
+  isDefaultTextNode(node: Node) {
+    return Element.isElementType(node, 'p')
+  },
+  isListNode(node: Node, type?: ListType) {
+    if (type) {
+      return Element.isElementType(node, type)
+    }
+    return (
+      Element.isElementType(node, 'ordered-list') ||
+      Element.isElementType(node, 'unordered-list')
+    )
+  },
+  isListItemNode(node: Node) {
+    return Element.isElementType(node, 'list-item')
+  },
+  isListItemTextNode(node: Node) {
+    return Element.isElementType(node, 'list-item-text')
+  },
+  createDefaultTextNode(props = {}) {
+    return { children: [{ text: '' }], ...props, type: 'p' } as Paragraph
+  },
+  createListNode(type: ListType = ListType.UNORDERED, props = {}) {
+    const nodeType =
+      type === ListType.ORDERED ? 'ordered-list' : 'unordered-list'
+    return { children: [{ text: '' }], ...props, type: nodeType } as
+      | OrderedList
+      | UnorderedList
+  },
+  createListItemNode(props = {}) {
+    return { children: [{ text: '' }], ...props, type: 'list-item' } as ListItem
+  },
+  createListItemTextNode(props = {}) {
+    return {
+      children: [{ text: '' }],
+      ...props,
+      type: 'list-item-child',
+    } as ListItemText
+  },
+})
